Update 8-3_go_fish.py
Function changes
pick_card(): changed method to get the index of the card picked by the player. New method is using a list comprehension to retrieve the index. the list created by the comprehension is then joined and stored as int-type in the variable "get_pick-index".

player_ask(): Added new variable "len_opponent_hand". This variable contains the length of the "opponent" list -1. This was originally used directly as condition in the second while loop. moved "pick_generator" to pick_card() function. Changed condition to check if player and opponent card are same value with call to function "same_value()". Variables "index" and "len_opponent_hand" are now reset to 0 and len(opponent)-1 when if condition is fulfilled and pairs have been removed.
Added "else:" after if condition to make sure the index is only incremented when the if condition isn't fulfilled.

General changes:
Removed "values_dict" redundant dictionary. The existing function "get_value()" returns the value of a card.