Update 8-3_go_fish.py
New Functions "player_go_fish()" and "computer_go_fish()". These are called by the player_ask() and computer_ask() functions when the card value picked by either the player or the computer has no matching value on the opponents hand. The Function prints a "Go Fish!" message, adds the top card of the deck to the hand and returns the hand.

Updated function "computer_pick()". Function now takes two arguments, "computer_hand" and "opponnent_last_picks". The function has a for each loop that loops over all values in "computer_hand" list, in the loop is an if-condition that checks if the card value is present in the "opponent_last_picks" list. If the condition is True, the value is returned and the loop is stopped. A random value picked from the "computer_hand" list is returned otherwise.
This update improves the intelligence of the "ai-player" it checks first if a previously requested card is in it's own hand and it requests this card.

updated player_ask() and computer_ask() functions: Both functions now return a list containing all pairs of that player as well as a list of containing all requests that where picked before of that user and didn't returna a match with the opponent.

Added empty "player_1_last_picks" and "player2_last_picks" lists in the game start code.

Added messages after a winning condition has been fulfilled. If a player wins, the length of both player tables is displayed and the winning player announced.