Update 9-2_go_fish.py
Implemented a "go fish" game where a human player plays against the computer.
Each player is dealt a hand of seven cards.
After the cards are dealt, each player checks his hand for pairs and moves those to his book.
The human player starts first and is asked to pick a card from his hand. If the card value matches a card value in the hand of the opposing player, both cards are removed and placed in the book of the human player. The human player is then asked to pick a new card from his hand.
If the card value has no match with the card values of the opposing player, the message to "Go Fish" is displayed and the player automatically takes the top card of the deck and adds it to his hand. If the card value of the added card matches a card value in the hand, both cards are removed and placed in the players book.

The computer always chooses a random card from his own hand, the card is then removed from his hand. If the card value of the chosen card matches with a card value from the human players hand, both cards are removed and placed in the computers book. If the card value has no match with the card values of the opposing player, then the chosen card is added back to the computers hand and the "Go Fish" message is displayed. The computer then adds the top card of the deck to his hand. The computer then checks for pairs just like the human player.