Update 9-3_scoring_go_fish.py
Renamed class "Hand" to class "Player".
changed removed default value for class attribute "label"  to positional argument.
added new attributes "self.book". This attribute is assigned to a variable in the driver code.
This was originally an empty list assigned to a variable in the driver code.
Game now tracks the card pairs for each player.