Create 9-4_ai_go_fish_w_classes.py
Created new script "9-4_ai_go_fish_w_classes.py"
This is a variation of script "9-3_scoring_go_fish". The computer remembers every card that has been asked by the player. The computer first checks if a card in it's hand was requested before by the player and picks the card from it's hand if there is a match. A random card is selected from the hand, when no card in the hand matches the previous request.