New file 10-1_go_fish_statistics.py
10-1_go_fish_statistics is an updated version of 9-4_ai_go_fish_w_classes.py.
The game now tracks statistics for each round played and appends them to a textfile "go_fish_statistics.txt".
Each time a game is finished, the time the game was finished, number of turns, winning and losing player as well as the number of pairs the winner and loser have are appended to the file.

To Do: implement a new function for tracking wins and losses of each player, write them to a different statistics file and display them at the start and end of each game.